MOO-152: Create DateUtils class
Move method getDateAtNDaysFromDate from service to this class.
